Resolve conflict in provider.go
