+++ internal/resources/fabric/connection/resource_schema.go
@@ -1,0 +1,657 @@
+package connection
+
+import (
+	equinix_fabric_schema "github.com/equinix/terraform-provider-equinix/internal/fabric/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
+)
+
+func fabricConnectionResourceSchema() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"type": {
+			Type:         schema.TypeString,
+			Required:     true,
+			ValidateFunc: validation.StringInSlice([]string{"EVPL_VC", "EPL_VC", "IP_VC", "IPWAN_VC", "ACCESS_EPL_VC", "EVPLAN_VC", "EPLAN_VC", "EIA_VC", "IA_VC", "EC_VC"}, false),
+			Description:  "Defines the connection type like EVPL_VC, EPL_VC, IPWAN_VC, IP_VC, ACCESS_EPL_VC, EVPLAN_VC, EPLAN_VC, EIA_VC, IA_VC, EC_VC",
+		},
+		"name": {
+			Type:         schema.TypeString,
+			Required:     true,
+			ValidateFunc: validation.StringLenBetween(1, 24),
+			Description:  "Connection name. An alpha-numeric 24 characters string which can include only hyphens and underscores",
+		},
+		"order": {
+			Type:        schema.TypeSet,
+			Optional:    true,
+			Computed:    true,
+			Description: "Order details",
+			MaxItems:    1,
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.OrderSch(),
+			},
+		},
+		"notifications": {
+			Type:        schema.TypeList,
+			Required:    true,
+			Description: "Preferences for notifications on connection configuration or status changes",
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.NotificationSch(),
+			},
+		},
+		"bandwidth": {
+			Type:        schema.TypeInt,
+			Required:    true,
+			Description: "Connection bandwidth in Mbps",
+		},
+		//"geo_scope": {
+		//	Type:         schema.TypeString,
+		//	Optional:     true,
+		//	ValidateFunc: validation.StringInSlice([]string{"CANADA", "CONUS"}, false),
+		//	Description:  "Geographic boundary types",
+		//},
+		"redundancy": {
+			Type:        schema.TypeSet,
+			Optional:    true,
+			Description: "Connection Redundancy Configuration",
+			MaxItems:    1,
+			Elem: &schema.Resource{
+				Schema: connectionRedundancySch(),
+			},
+		},
+		"a_side": {
+			Type:        schema.TypeSet,
+			Required:    true,
+			Description: "Requester or Customer side connection configuration object of the multi-segment connection",
+			MaxItems:    1,
+			Elem:        connectionSideSch(),
+			Set:         schema.HashResource(accessPointSch()),
+		},
+		"z_side": {
+			Type:        schema.TypeSet,
+			Required:    true,
+			Description: "Destination or Provider side connection configuration object of the multi-segment connection",
+			MaxItems:    1,
+			Elem:        connectionSideSch(),
+			Set:         schema.HashResource(accessPointSch()),
+		},
+		"project": {
+			Type:        schema.TypeSet,
+			Optional:    true,
+			Computed:    true,
+			Description: "Project information",
+			MaxItems:    1,
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.ProjectSch(),
+			},
+		},
+		"additional_info": {
+			Type:        schema.TypeList,
+			Optional:    true,
+			Description: "Connection additional information",
+			Elem: &schema.Schema{
+				Type: schema.TypeMap,
+			},
+		},
+		"href": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Connection URI information",
+		},
+		"uuid": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Equinix-assigned connection identifier",
+		},
+		"description": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Customer-provided connection description",
+		},
+		"state": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Connection overall state",
+		},
+		"operation": {
+			Type:        schema.TypeSet,
+			Computed:    true,
+			Description: "Connection type-specific operational data",
+			Elem: &schema.Resource{
+				Schema: operationSch(),
+			},
+		},
+		"account": {
+			Type:        schema.TypeSet,
+			Computed:    true,
+			Description: "Customer account information that is associated with this connection",
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.AccountSch(),
+			},
+		},
+		"change_log": {
+			Type:        schema.TypeSet,
+			Computed:    true,
+			Description: "Captures connection lifecycle change information",
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.ChangeLogSch(),
+			},
+		},
+		"is_remote": {
+			Type:        schema.TypeBool,
+			Computed:    true,
+			Description: "Connection property derived from access point locations",
+		},
+		"direction": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Connection directionality from the requester point of view",
+		},
+	}
+}
+
+func connectionSideSch() *schema.Resource {
+	return &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"service_token": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "For service token based connections, Service tokens authorize users to access protected resources and services. Resource owners can distribute the tokens to trusted partners and vendors, allowing selected third parties to work directly with Equinix network assets",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: serviceTokenSch(),
+				},
+			},
+			"access_point": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "Point of access details",
+				MaxItems:    1,
+				Elem:        accessPointSch(),
+			},
+			"additional_info": {
+				Type:        schema.TypeList,
+				Optional:    true,
+				Description: "Connection side additional information",
+				Elem: &schema.Resource{
+					Schema: additionalInfoSch(),
+				},
+			},
+		},
+	}
+}
+
+func serviceTokenSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"type": {
+			Type:         schema.TypeString,
+			Optional:     true,
+			ValidateFunc: validation.StringInSlice([]string{"VC_TOKEN"}, true),
+			Description:  "Token type - VC_TOKEN",
+		},
+		"href": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "An absolute URL that is the subject of the link's context",
+		},
+		"uuid": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Equinix-assigned service token identifier",
+		},
+		"description": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Service token description",
+		},
+	}
+}
+
+func accessPointSch() *schema.Resource {
+	return &schema.Resource{
+		Schema: map[string]*schema.Schema{
+			"type": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				ValidateFunc: validation.StringInSlice([]string{"COLO", "VD", "VG", "SP", "IGW", "SUBNET", "CLOUD_ROUTER", "NETWORK", "METAL_NETWORK"}, true),
+				Description:  "Access point type - COLO, VD, VG, SP, IGW, SUBNET, CLOUD_ROUTER, NETWORK",
+			},
+			"account": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Computed:    true,
+				Description: "Account",
+				Elem: &schema.Resource{
+					Schema: equinix_fabric_schema.AccountSch(),
+				},
+			},
+			"location": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Computed:    true,
+				Description: "Access point location",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: equinix_fabric_schema.LocationSch(),
+				},
+			},
+			"port": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "Port access point information",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: portSch(),
+				},
+			},
+			"profile": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "Service Profile",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: serviceProfileSch(),
+				},
+			},
+			"gateway": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Deprecated:  "use router attribute instead; gateway is no longer a part of the supported backend",
+				Description: "**Deprecated** `gateway` Use `router` attribute instead",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: cloudRouterSch(),
+				},
+			},
+			"router": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "Cloud Router access point information that replaces `gateway`",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: cloudRouterSch(),
+				},
+			},
+			"link_protocol": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "Connection link protocol",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: accessPointLinkProtocolSch(),
+				},
+			},
+			"virtual_device": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "Virtual device",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: accessPointVirtualDeviceSch(),
+				},
+			},
+			"interface": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "Virtual device interface",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: accessPointInterface(),
+				},
+			},
+			"network": {
+				Type:        schema.TypeSet,
+				Optional:    true,
+				Description: "network access point information",
+				MaxItems:    1,
+				Elem: &schema.Resource{
+					Schema: networkSch(),
+				},
+			},
+			"seller_region": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Access point seller region",
+			},
+			"peering_type": {
+				Type:         schema.TypeString,
+				Optional:     true,
+				ValidateFunc: validation.StringInSlice([]string{"PRIVATE", "MICROSOFT", "PUBLIC", "MANUAL"}, true),
+				Description:  "Peering Type- PRIVATE,MICROSOFT,PUBLIC, MANUAL",
+			},
+			"authentication_key": {
+				Type:        schema.TypeString,
+				Optional:    true,
+				Description: "Authentication key for provider based connections",
+			},
+			"provider_connection_id": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "Provider assigned Connection Id",
+			},
+		},
+	}
+}
+
+func serviceProfileSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"href": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Service Profile URI response attribute",
+		},
+		"type": {
+			Type:         schema.TypeString,
+			Required:     true,
+			ValidateFunc: validation.StringInSlice([]string{"L2_PROFILE", "L3_PROFILE", "ECIA_PROFILE", "ECMC_PROFILE", "IA_PROFILE"}, true),
+			Description:  "Service profile type - L2_PROFILE, L3_PROFILE, ECIA_PROFILE, ECMC_PROFILE, IA_PROFILE",
+		},
+		"name": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Customer-assigned service profile name",
+		},
+		"uuid": {
+			Type:        schema.TypeString,
+			Required:    true,
+			Description: "Equinix assigned service profile identifier",
+		},
+		"description": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "User-provided service description",
+		},
+		"access_point_type_configs": {
+			Type:        schema.TypeList,
+			Computed:    true,
+			Description: "Access point config information",
+			Elem: &schema.Resource{
+				Schema: connectionAccessPointTypeConfigSch(),
+			},
+		},
+	}
+}
+
+func cloudRouterSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"uuid": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Equinix-assigned virtual gateway identifier",
+		},
+		"href": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Unique Resource Identifier",
+		},
+	}
+}
+
+func accessPointLinkProtocolSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"type": {
+			Type:         schema.TypeString,
+			Optional:     true,
+			Description:  "Type of the link protocol - UNTAGGED, DOT1Q, QINQ, EVPN_VXLAN",
+			ValidateFunc: validation.StringInSlice([]string{"UNTAGGED", "DOT1Q", "QINQ", "EVPN_VXLAN"}, true),
+		},
+		"vlan_tag": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "Vlan Tag information, vlanTag value specified for DOT1Q connections",
+		},
+		"vlan_s_tag": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "Vlan Provider Tag information, vlanSTag value specified for QINQ connections",
+		},
+		"vlan_c_tag": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Computed:    true,
+			Description: "Vlan Customer Tag information, vlanCTag value specified for QINQ connections",
+		},
+	}
+}
+
+func accessPointVirtualDeviceSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"href": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Unique Resource Identifier",
+		},
+		"uuid": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Equinix-assigned Virtual Device identifier",
+		},
+		"type": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Virtual Device type",
+		},
+		"name": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Customer-assigned Virtual Device Name",
+		},
+	}
+}
+
+func accessPointInterface() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"uuid": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Equinix-assigned interface identifier",
+		},
+		"id": {
+			Type:        schema.TypeInt,
+			Computed:    true,
+			Optional:    true,
+			Description: "id",
+		},
+		"type": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Interface type",
+		},
+	}
+}
+
+func networkSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"uuid": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Equinix-assigned Network identifier",
+		},
+		"href": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Unique Resource Identifier",
+		},
+	}
+}
+
+func portRedundancySch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"enabled": {
+			Type:        schema.TypeBool,
+			Computed:    true,
+			Description: "Access point redundancy",
+		},
+		"group": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Port redundancy group",
+		},
+		"priority": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Priority type-Primary or Secondary",
+		},
+	}
+}
+
+func portSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"uuid": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Computed:    true,
+			Description: "Equinix-assigned Port identifier",
+		},
+		"href": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Unique Resource Identifier",
+		},
+		"name": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Port name",
+		},
+		"redundancy": {
+			Type:        schema.TypeSet,
+			Computed:    true,
+			Description: "Redundancy Information",
+			Elem: &schema.Resource{
+				Schema: portRedundancySch(),
+			},
+		},
+	}
+}
+
+func connectionAccessPointTypeConfigSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"type": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Type of access point type config - VD, COLO",
+		},
+		"uuid": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Equinix-assigned access point type config identifier",
+		},
+	}
+}
+
+func additionalInfoSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"key": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Additional information key",
+		},
+		"value": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Additional information value",
+		},
+	}
+}
+
+func operationSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"provider_status": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Connection provider readiness status",
+		},
+		"equinix_status": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Description: "Connection status",
+		},
+		"errors": {
+			Type:        schema.TypeList,
+			Computed:    true,
+			Description: "Errors occurred",
+			Elem: &schema.Resource{
+				Schema: equinix_fabric_schema.ErrorSch(),
+			},
+		},
+	}
+}
+
+func connectionRedundancySch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"group": {
+			Type:        schema.TypeString,
+			Computed:    true,
+			Optional:    true,
+			Description: "Redundancy group identifier (Use the redundancy.0.group UUID of primary connection; e.g. one(equinix_fabric_connection.primary_port_connection.redundancy).group or equinix_fabric_connection.primary_port_connection.redundancy.0.group)",
+		},
+		"priority": {
+			Type:         schema.TypeString,
+			Computed:     true,
+			Optional:     true,
+			ValidateFunc: validation.StringInSlice([]string{"PRIMARY", "SECONDARY"}, true),
+			Description:  "Connection priority in redundancy group - PRIMARY, SECONDARY",
+		},
+	}
+}
+
+func createApiConfigSch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"api_available": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Indicates if it's possible to establish connections based on the given service profile using the Equinix Fabric API.",
+		},
+		"equinix_managed_vlan": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Setting indicating that the VLAN is managed by Equinix (true) or not (false)",
+		},
+		"allow_over_subscription": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Setting showing that oversubscription support is available (true) or not (false). The default is false",
+		},
+		"over_subscription_limit": {
+			Type:        schema.TypeInt,
+			Optional:    true,
+			Description: "Port bandwidth multiplier that determines the total bandwidth that can be allocated to users creating connections to your services. For example, a 10 Gbps port combined with an overSubscriptionLimit parameter value of 10 allows your subscribers to create connections with a total bandwidth of 100 Gbps.",
+		},
+		"bandwidth_from_api": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Indicates if the connection bandwidth can be obtained directly from the cloud service provider.",
+		},
+		"integration_id": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "A unique identifier issued during onboarding and used to integrate the customer's service profile with the Equinix Fabric API.",
+		},
+		"equinix_managed_port": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Setting indicating that the port is managed by Equinix (true) or not (false)",
+		},
+	}
+}
+
+func createAuthenticationKeySch() map[string]*schema.Schema {
+	return map[string]*schema.Schema{
+		"required": {
+			Type:        schema.TypeBool,
+			Optional:    true,
+			Description: "Requirement to configure an authentication key.",
+		},
+		"label": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Name of the parameter that must be provided to authorize the connection.",
+		},
+		"description": {
+			Type:        schema.TypeString,
+			Optional:    true,
+			Description: "Description of authorization key",
+		},
+	}
+}
